Experimental use of Werkzeug serve as alternative to apistar
This commit includes a new `run2` method for TFServeApp, which uses
Werkzeug. The existing `run` method is unchanged to maintain
compatibility with the current API and behavior.

`run2` insulates the programmer from the Werkzeug interface with the
exception of an optional middleware argument that can be used by the
programmer to customize the HTTP request/response behavior (this is
used by `main.py`, which will land in a forthcoming commit).

`test_serve.py` tests the major features of the implementation and
illustrates its use.