main module and tfserve console script support
`main.py` implements a basic script that can be used to run TFServe
from a console. The module can be used as a program by running:

    python -m tfserve.main

The module supports the `--help` option, which shows command help.

`setup.py` has been modified to include `tfserve` as a console script
using entry points, which is the standard method of installing Python
console scripts. When `tfserve` is installed using pip, an executable
script `tfserve` is created in the environment binary directory. This
script is removed when the package is uninstalled.

`test_main.py` tests the main module implementation and illustrates
its use.