Check for tensorflow on package import
If the package is imported as a result of a tfserve script invocation,
it prints a message and exits with an error code. Otherwise any
tensorflow import errors are ignored.

This has the effect of giving a console script user a error message
and advice for resolving the issue. Programmers who are attempting to
import tfserve are not effect by this behavior - if tensorflow is not
installed, Python will raise an import error as it does normally.